pyproject.toml
[tool.blue]
line-length = 79  # this is the default
[tool.isort]
ensure_newline_before_comments = true
line_length = 79
multi_line_output = 3
include_trailing_comma = true
skip_glob = ["docs/*", "*.egg-info"]

[tool.poetry]
name = "conda-on-air"
version = "0.0.1" # semantic-release
description = "A project that allows the local build and installation of an existent conda-forge feedstock"
authors = ["Ivan Ogasawara <ivan.ogasawara@gmail.com>"]
license = "BSD 3 Clause"
include = [
  "*.cfg",
  "*.ini",
  "*.js",
  "*.json",
  "*.lock",
  "*.md",
  "*.py",
  "*.sh",
  "*.rst",
  "*.txt",
  "*.toml",
  "*.yml",
  "*.yaml",
  ".github",
  "conda",
  "docker",
  "docs",
  ".dockerignore",
  ".env.tpl",
  ".gitignore",
  "Makefile",
]
exclude = [
  ".git/*",
  ".env*",
]

[tool.poetry.dependencies]
python = "^3.8.1"
pyyaml = "^6.0"
urllib3 = "<2"
sh = "^2.0.4"

[tool.poetry.dev-dependencies]
pytest = "^7"
blue = "^0.9.1"

isort = "^5"
pre-commit = "2"
flake8 = "^4"
mypy = "^0.931"
cookiecutter = "^1.7.3"
pytest-cov = "^3.0.0"
Jinja2 = "<3.1.0"
mkdocs = ">=1.2.3,<2"
mkdocs-exclude = ">=1.0.2,<2"
mkdocs-jupyter = ">=0.20.0,<1"
mkdocs-literate-nav = ">=0.4.1,<1"
mkdocs-macros-plugin = ">=0.6.3,<1"
mkdocs-material = ">=8.2.1,<9"
mkdocstrings = ">=0.17.0,<0.18.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
